Fix parsing of the output for ip -4 route
The output of `ip -4 route` is not parsed correctly.

This patch makes the parsing less wrong, but should probably be revised
further. The problem is that lines like this exists:

  172.17.0.0/16 dev docker0 proto kernel scope link src 172.17.0.1 linkdown

This is not of the format r"target (key values)*", because the
"linkdown" does not have a value. I am not sure what to do about it. For
now we just assume that any such fields are unimportant and comes after
any important ones.